+++ src/OptionsResolver/PaymentMethodsSettingsResolver.php
@@ -52,6 +52,7 @@ class PaymentMethodsSettingsResolver extends AbstractSettingsResolver
                 'type',
                 'logo',
                 'deleteLogo',
+                'fixedSoftDescriptor',
             ])
             ->setNormalizer(
                 'displayGenericOption',