+++ translations/fr.php
@@ -141,6 +141,8 @@ $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_254ad5cf512e99709dbcb
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_ff76e4c4c9abdc7de325c97912ffb8cd'] = 'Nous vous recommandons d’utiliser des images ayant une hauteur maximale de 20 px et une longueur maximale de 120 px';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_3c8990c77ed9e1634b5cdd133c6e0993'] = 'Titre du bouton de paiement';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_6f1ad8444a503e36d821da11b4b1fe57'] = 'Titre du bouton de sélection de paiement sur votre page de passage de commande';
+$_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_46310871a02a521ecfdfe034d1fd0085'] = 'Libellé de facturation';
+$_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_0b9f4e7ca03c57b2d6943e035e6c996e'] = 'Minimum 2, maximum 15 caractères. Le libellé de facturation, aussi appelé Soft descriptor, vous permet de communiquer une valeur aux banques traitant le paiement. Si la banque émettrice le permet, elle affichera cette référence sur le relevé bancaire de l\'utilisateur. Note importante : renseigner ce champ est obligatoire pour pouvoir utiliser la méthode de paiement Pledg.';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_55e58b9f475a7a98a03b777f4e0a2ec7'] = 'Sélection de la méthode de paiement avant redirection';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_df2a60fbc299477de52630ae7fe6e477'] = 'Chaque méthode de paiement est identifiée à l’aide d’un bouton. Lorsqu’il clique, le client est redirigé pour payer sur la page de paiement spécifique à la marque';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_6ef52d9cbb060e8e35b86085db5536cb'] = 'Rafraîchir la liste des méthodes de paiement disponibles';