+++ translations/it.php
@@ -141,6 +141,8 @@ $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_254ad5cf512e99709dbcb
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_ff76e4c4c9abdc7de325c97912ffb8cd'] = 'Si consiglia di utilizzare immagini di massimo 20 px di altezza e 120 px di lunghezza';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_3c8990c77ed9e1634b5cdd133c6e0993'] = 'Titolo del tasto di pagamento';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_6f1ad8444a503e36d821da11b4b1fe57'] = 'Titolo del pulsante di selezione dei pagamenti sulla pagina di check-out';
+$_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_46310871a02a521ecfdfe034d1fd0085'] = 'Descrizione fissa';
+$_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_0b9f4e7ca03c57b2d6943e035e6c996e'] = 'Minimo 2, massimo 15 caratteri. È possibile specificare un valore fisso che sarà inviato alle banche che gestiscono il pagamento. Se la banca del consumatore supporta questa funzione, questo valore può essere visualizzato sull’estratto conto bancario del consumatore. Nota importante: impostare un valore qui è obbligatorio per utilizzare il metodo di pagamento Pledg.';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_55e58b9f475a7a98a03b777f4e0a2ec7'] = 'Selezione del metodo di pagamento prima del reindirizzamento';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_df2a60fbc299477de52630ae7fe6e477'] = 'Ciascun metodo di pagamento è identificato da un tasto. Con un clic, il cliente viene reindirizzato al pagamento sulla pagina di pagamento specifica del marchio';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_6ef52d9cbb060e8e35b86085db5536cb'] = 'Aggiornare l’elenco di metodi di pagamento disponibili';