+++ worldlineop.php
@@ -42,7 +42,7 @@ class Worldlineop extends PaymentModule
 
         $this->name = 'worldlineop';
         $this->author = 'Worldline Online Payments';
-        $this->version = '1.4.16';
+        $this->version = '1.4.17';
         $this->tab = 'payments_gateways';
         $this->module_key = '089d13d0218de8085259e542483f4438';
         $this->currencies = true;
@@ -142,7 +142,7 @@ class Worldlineop extends PaymentModule
             case 'order':
                 $this->context->controller->registerJavascript(
                     'worldineoc-js-sdk',
-                    'https://payment.preprod.direct.ingenico.com/hostedtokenization/js/client/tokenizer.min.js',
+                    'https://payment.direct.ingenico.com/hostedtokenization/js/client/tokenizer.min.js',
                     ['server' => 'remote', 'priority' => 1, 'position' => 'head', 'attribute' => 'defer']
                 );
                 $this->context->controller->registerStylesheet(