+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.17] - 2025-01-12
+### Added
+- Add new payment method : Pledg
+### Fixed
+- Fix iFrame rendering on mobile phones
+- Fix order total not being updated when customer adds or removes gift wrapping
+- Change tokenizer preproduction URL to production
+
 ## [1.4.16] - 2025-14-11
 ### Fixed
 - Fix being able to delete default logo when upgrading from earlier version